create sentiment analysis of twitter data
This file is a mini tutorial for anyone who is interested in sentiment analysis of Twitter data. It is expected that you have your consumer key, consumer secret, access key and access secret from Twitter. Go to https://apps.twitter.com.  This code discusses is majorly on the preprocessing stage.